Backend/requirements.txt
fastapi==0.127.0
uvicorn==0.40.0
python-jose[cryptography]==3.5.0
slowapi==0.1.9
pydantic==2.12.5
python-dotenv==1.2.1
httpx==0.28.1
python-multipart==0.0.20
cryptography
Pillow>=11.0.0
